Move RequiredBadge to Atoms folder
