+++ src/components/atoms/RequiredBadge/const.ts
